+++ src/IntegrationTests/RequestFactoringSettingApplyBehavior.cs
@@ -49,7 +49,7 @@ namespace IntegrationTests
         [Theory]
         [InlineData(true)]
         [InlineData(false)]
-        public async Task ShouldApplyIgnoreNullFields(bool ignore)
+        public async Task ShouldApplyIgnoreJsonNullFields(bool ignore)
         {
             //Arrange
             var model = new TestModel
@@ -60,7 +60,7 @@ namespace IntegrationTests
             var proxy = CreateProxy(o => o.JsonSettings.IgnoreNullValues = ignore);
 
             //Act
-            var resp = await proxy.CallEchoObj(model);
+            var resp = await proxy.CallEchoJson(model);
             
             //Assert
             if(ignore)
@@ -70,7 +70,7 @@ namespace IntegrationTests
         }
 
         [Fact]
-        public async Task ShouldIgnoreNullFieldsByDefault()
+        public async Task ShouldIgnoreJsonNullFieldsByDefault()
         {
             //Arrange
             var model = new TestModel
@@ -81,17 +81,96 @@ namespace IntegrationTests
             var proxy = CreateProxy(o =>{});
 
             //Act
-            var resp = await proxy.CallEchoObj(model);
+            var resp = await proxy.CallEchoJson(model);
 
             //Assert
             Assert.DoesNotContain("\"TestValue\":null", resp);
         }
 
+        [Theory]
+        [InlineData(true)]
+        [InlineData(false)]
+        public async Task ShouldApplyIgnoreUrlFormNullFields(bool ignore)
+        {
+            //Arrange
+            var model = new TestModel
+            {
+                TestValue = null
+            };
+
+            var proxy = CreateProxy(o => o.UrlFormSettings.IgnoreNullValues = ignore);
+
+            //Act
+            var resp = await proxy.CallEchoForm(model);
+
+            //Assert
+            Assert.Equal(ignore ? null : "TestValue=", resp);
+        }
+
+        [Fact]
+        public async Task ShouldIgnoreUrlFormNullFieldsByDefault()
+        {
+            //Arrange
+            var model = new TestModel
+            {
+                TestValue = null
+            };
+
+            var proxy = CreateProxy(o => { });
+
+            //Act
+            var resp = await proxy.CallEchoForm(model);
+
+            //Assert
+            Assert.Equal("TestValue=", resp);
+        }
+
+        [Theory]
+        [InlineData(true)]
+        [InlineData(false)]
+        public async Task ShouldApplyFormEscape(bool escape)
+        {
+            //Arrange
+            var model = new TestModel
+            {
+                TestValue = "http://test.com/"
+            };
+
+            var proxy = CreateProxy(o => o.UrlFormSettings.EscapeSymbols = escape);
+
+            //Act
+            var resp = await proxy.CallEchoForm(model);
+
+            //Assert
+            Assert.Contains(escape ? "http%3A%2F%2Ftest.com%2F" : "http://test.com/", resp);
+        }
+
+        [Fact]
+        public async Task ShouldEscapeFormByDefault()
+        {
+            //Arrange
+            var model = new TestModel
+            {
+                TestValue = "http://test.com/"
+            };
+
+            var proxy = CreateProxy(o => { });
+
+            //Act
+            var resp = await proxy.CallEchoForm(model);
+
+            //Assert
+            Assert.Contains("http%3A%2F%2Ftest.com%2F", resp);
+        }
+
         [Api(Key = "No matter for this test")]
         interface ITestServer
         {
-            [Get("echo/obj")]
-            Task<string> CallEchoObj([JsonContent] TestModel msg);
+            [Get("echo/body")]
+            Task<string> CallEchoJson([JsonContent] TestModel msg);
+
+            [Get("echo/body")]
+            Task<string> CallEchoForm([FormContent] TestModel msg);
         }
     }
 }