+++ src/MyLab.ApiClient/RequestFactoringSettings.cs
@@ -13,14 +13,20 @@ namespace MyLab.ApiClient
         public JsonSerializerSettings JsonSettings { get; set; }
 
         /// <summary>
+        /// Defines url-encoded-form serialization settings
+        /// </summary>
+        public ApiUrlFormSettings UrlFormSettings { get; set; }
+
+
+        /// <summary>
         /// Creates req factoring settings from app options
         /// </summary>
         /// <param name="options"></param>
         /// <returns></returns>
-        public static RequestFactoringSettings CreateFromOptions(ApiClientsOptions options)
+        public static RequestFactoringSettings CreateFromOptions(ApiClientsOptions? options)
         {
-            var apiJsonSettings = options.JsonSettings ?? new ApiJsonSettings();
-
+            var apiJsonSettings = options?.JsonSettings ?? new ApiJsonSettings();
+            
             var jsonSerializerSettings = new JsonSerializerSettings
             {
                 NullValueHandling = apiJsonSettings.IgnoreNullValues
@@ -30,8 +36,9 @@ namespace MyLab.ApiClient
 
             return new RequestFactoringSettings
             {
-                JsonSettings = jsonSerializerSettings
-            };
+                JsonSettings = jsonSerializerSettings,
+                UrlFormSettings = options?.UrlFormSettings ?? new ApiUrlFormSettings()
+        };
         }
     }
 }